refactor(cli): move inline tests to *_tests.rs siblings
Extract the inline test modules from license.rs and main.rs into sibling
license_tests.rs / main_tests.rs files (wired via #[cfg(test)] #[path]),
matching the project's *_tests.rs convention. Test files are excluded from
Codacy, so this drops each source file's measured complexity without
touching any production logic. All 26 tests run and pass unchanged.
